#1044c: split cli.go (2,661 → 1,952 LOC) — last #1044 file under threshold
Phase 3 of #1044 (modularity-discipline cleanup). Closes the audit
that started with #1044a (compiler.go shipped via #1092: 3,509 →
1,663 LOC) and #1044b (daemon.go shipped via #1091: 2,851 → 346 LOC).

Move four large handler functions from `pkg/cli/cli.go` into existing
sibling files in `pkg/cli/`. Pure relocation — same methodology as the
just-completed #1043 server_show.go split. No new files, no
behavior change.

| Function           | LOC | Target file (already existed) |
|--------------------|-----|-------------------------------|
| handleShow         | 215 | cli_show.go                   |
| handleShowSystem   | 209 | cli_show_system.go            |
| valueProvider      | 166 | completion.go                 |
| handleConfigShow   | 124 | cli_config.go                 |

`handleShowSecurity` (251 LOC) is intentionally LEFT in cli.go: its
natural target `cli_show_security.go` was already 1,978 LOC before
this PR; appending the dispatcher would push that file over the same
threshold. Splitting `cli_show_security.go` further is a separate
follow-up.

LOC after this PR:
- pkg/cli/cli.go:               2,661 → 1,952  (-714)
- pkg/cli/cli_show.go:              0 →   228  (new)
- pkg/cli/cli_show_security.go: 1,978 → 1,978  (unchanged)
- pkg/cli/cli_show_system.go:     768 →   977  (+209)
- pkg/cli/completion.go:          146 →   313  (+167)
- pkg/cli/cli_config.go:          323 →   447  (+124)

The `cmdtree` import migrates to cli_show.go (cli.go retains its
own cmdtree references for other call sites).

All 880+ Go tests pass.

Refs #1044.
